Pyramid TIFF: expand type checking to look at image dimensions
The presence of a Software tag containing "Faas" is now sufficient but
not necessary for a TIFF file to be detected as pyramid TIFF.  If a
valid Software tag is not present, the width and height of the first and
last IFDs will be compared.  If the last IFD's dimensions are the same
power of 2 smaller than the first IFD's dimensions, then the file is
assumed to be pyramid TIFF.

As this can affect performance of files that require readers listed after
PyramidTiffReader in readers.txt, only the first and last IFDs are checked,
but it would be possible to check all of them and/or require more than
two IFDs.