Analyze: change pixelsFilename assumptions
pixelsFilename is now assumed to be a relative file path, where
previously it was an absolute path.  Memo files written prior to this
commit will have the absolute path cached.  Reading the old memo files
with this commit will result in invalidation, since the pixels file path
on reopenFile will be incorrect.  Reading memo files written with this
commit will work, since the relative path will have been cached.

Other readers will not be affected since the type and count of the
reader's variables remains unchanged.